third_party/WebKit/LayoutTests/netinfo/type-change-no-listener.html

<!DOCTYPE html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/netinfo_common.js"></script>
</head>
<body>
<script>
description('Tests that the type changes without an active listener');

internals.setNetworkConnectionInfoOverride(isTypeOnline(newConnectionType), newConnectionType, newEffectiveType, newRtt, newDownlinkMax);
shouldBe('connection.type', 'newConnectionType');
shouldBe('connection.downlinkMax', 'newDownlinkMax');
shouldBe('connection.effectiveType', 'newEffectiveType');
shouldBe('connection.rtt', 'newRtt', false, newRttMaxNoise);
shouldBe('connection.downlink', 'newDownlinkMax', false, newDownlinkMaxNoise);

finishJSTest();
</script>
</body>
</html>
